renaming list handler
