+++ cmd/web/api/json/handlers/savings_goals/entrypoint.go
@@ -1,9 +1,9 @@
 package savings_goals
 
 import (
-	"financo/cmd/web/api/json/handlers/savings_goals/active_goals_for_handler"
 	"financo/cmd/web/api/json/handlers/savings_goals/create_handler"
 	"financo/cmd/web/api/json/handlers/savings_goals/delete_handler"
+	"financo/cmd/web/api/json/handlers/savings_goals/list_handler"
 	"financo/cmd/web/api/json/handlers/savings_goals/mark_as_achieved_handler"
 	"financo/cmd/web/api/json/handlers/savings_goals/reorder_handler"
 	"financo/cmd/web/api/json/handlers/savings_goals/update_handler"
@@ -12,13 +12,13 @@ import (
 )
 
 func Routes(r chi.Router) {
+	r.Get("/", list_handler.HandlerFunc)
 	r.Post("/", create_handler.HandlerFunc)
-	r.Get("/", active_goals_for_handler.HandlerFunc)
-	r.Patch("/reorder", reorder_handler.HandlerFunc)
+	r.Put("/reorder", reorder_handler.HandlerFunc)
 
-	r.Route("/{id}", func(r chi.Router) {
-		r.Delete("/", delete_handler.HandlerFunc)
-		r.Put("/", update_handler.HandlerFunc)
-		r.Patch("/mark-as-achieved", mark_as_achieved_handler.HandlerFunc)
+	r.Route("/{id}", func(goal chi.Router) {
+		goal.Delete("/", delete_handler.HandlerFunc)
+		goal.Put("/", update_handler.HandlerFunc)
+		goal.Patch("/mark-as-achieved", mark_as_achieved_handler.HandlerFunc)
 	})
 }
